Conflicts:
	fiber/__init__.py

+++ fiber/admin.py
@@ -4,14 +4,16 @@ from django.contrib.admin.util import model_ngettext
 from django.utils.translation import ugettext_lazy as _
 from django.core.exceptions import PermissionDenied
 
-from multilingual.admin.options import MultilingualModelAdmin
+from hvad.admin import TranslatableAdmin
 
 from . import admin_forms as forms
 from . import fiber_admin
-from .app_settings import TEMPLATE_CHOICES, CONTENT_TEMPLATE_CHOICES, PERMISSION_CLASS
+from .app_settings import TEMPLATE_CHOICES, CONTENT_TEMPLATE_CHOICES, PERMISSION_CLASS, IMAGE_PREVIEW
 from .editor import get_editor_field_name
+from .fiber_admin import ModelAdmin
 from .models import Page, ContentItem, PageContentItem, Image, File
 from .utils.class_loader import load_class
+from .utils.widgets import AdminImageWidgetWithPreview
 
 perms = load_class(PERMISSION_CLASS)
 
@@ -48,15 +50,16 @@ class UserPermissionMixin(object):
         perms.object_created(request.user, obj)
 
 
-class FileAdmin(UserPermissionMixin, MultilingualModelAdmin):
-    list_display = ('__unicode__', 'title', )
+class FileAdmin(UserPermissionMixin, ModelAdmin):
+    list_display = ('__unicode__', 'title', 'updated',)
     date_hierarchy = 'updated'
     search_fields = ('title', )
     actions = ['really_delete_selected']
 
     def get_actions(self, request):
         actions = super(FileAdmin, self).get_actions(request)
-        del actions['delete_selected']  # the original delete selected action doesn't remove associated files, because .delete() is never called
+        if 'delete_selected' in actions:
+            del actions['delete_selected']  # the original delete selected action doesn't remove associated files, because .delete() is never called
         return actions
 
     def really_delete_selected(self, request, queryset):
@@ -76,11 +79,26 @@ class FileAdmin(UserPermissionMixin, MultilingualModelAdmin):
 
 
 class ImageAdmin(FileAdmin):
-    pass
+    list_display = ('__unicode__', 'title', 'get_size', 'updated',)
+    fieldsets = (
+        (None, {'fields': ('image', 'title',)}),
+        (_('Size'), {'classes': ('collapse',), 'fields': ('width', 'height',)}),
+    )
 
 
-class ContentItemAdmin(UserPermissionMixin, MultilingualModelAdmin):
-    list_display = ('__unicode__',)
+class ImageAdminWithPreview(ImageAdmin):
+    list_display = ('preview', '__unicode__', 'title', 'get_size', 'updated',)
+
+    def formfield_for_dbfield(self, db_field, **kwargs):
+        if db_field.name == 'image':
+            request = kwargs.pop("request", None)
+            kwargs['widget'] = AdminImageWidgetWithPreview
+            return db_field.formfield(**kwargs)
+        return super(ImageAdminWithPreview, self).formfield_for_dbfield(db_field, **kwargs)
+
+
+class ContentItemAdmin(UserPermissionMixin, TranslatableAdmin):
+    list_display = ('__unicode__', 'unused')
     form = forms.ContentItemAdminForm
     fieldsets = (
         (None, {'fields': ('name', get_editor_field_name('content_html'),)}),
@@ -90,13 +108,19 @@ class ContentItemAdmin(UserPermissionMixin, MultilingualModelAdmin):
     date_hierarchy = 'updated'
     search_fields = ('name', get_editor_field_name('content_html'))
 
+    def unused(self, obj):
+        if obj.used_on_pages_data is None:
+            return True
+        return False
+    unused.boolean = True
+
 
 class PageContentItemInline(UserPermissionMixin, admin.TabularInline):
     model = PageContentItem
     extra = 1
 
 
-class PageAdmin(UserPermissionMixin, MultilingualModelAdmin):
+class PageAdmin(UserPermissionMixin, ModelAdmin):
 
     form = forms.PageForm
     fieldsets = (
@@ -106,7 +130,7 @@ class PageAdmin(UserPermissionMixin, MultilingualModelAdmin):
     )
 
     inlines = (PageContentItemInline,)
-    list_display = ('title', 'view_on_site', 'url', 'redirect_page', 'get_absolute_url', 'action_links')
+    list_display = ('__unicode__', 'view_on_site', 'url', 'redirect_page', 'get_absolute_url', 'action_links')
     list_per_page = 1000
     search_fields = ('title', 'url', 'redirect_page__title')
 
@@ -147,7 +171,7 @@ class PageAdmin(UserPermissionMixin, MultilingualModelAdmin):
     action_links.allow_tags = True
 
 
-class FiberAdminContentItemAdmin(UserPermissionMixin, MultilingualModelAdmin):
+class FiberAdminContentItemAdmin(UserPermissionMixin, TranslatableAdmin):
     list_display = ('__unicode__',)
     form = forms.ContentItemAdminForm
 
@@ -165,7 +189,7 @@ class FiberAdminContentItemAdmin(UserPermissionMixin, MultilingualModelAdmin):
             )
 
 
-class FiberAdminPageAdmin(UserPermissionMixin, MultilingualModelAdmin):
+class FiberAdminPageAdmin(UserPermissionMixin, ModelAdmin):
 
     form = forms.PageForm
 
@@ -177,13 +201,13 @@ class FiberAdminPageAdmin(UserPermissionMixin, MultilingualModelAdmin):
             self.fieldsets = (
                 (None, {'fields': ('title', 'url', )}),
                 (_('Advanced options'), {'fields': ('redirect_page', 'show_in_menu', 'is_public', )}),
-                (_('SEO'), {'fields': ('meta_description', )}),
+                (_('SEO'), {'fields': ('doc_title', 'meta_description', 'meta_keywords', )}),
             )
         else:
             self.fieldsets = (
                 (None, {'fields': ('title', 'url', )}),
                 (_('Advanced options'), {'fields': ('template_name', 'redirect_page', 'show_in_menu', 'is_public', )}),
-                (_('SEO'), {'fields': ('meta_description', )}),
+                (_('SEO'), {'fields': ('doc_title', 'meta_description', 'meta_keywords', )}),
             )
 
     def save_model(self, request, obj, form, change):
@@ -204,7 +228,11 @@ class FiberAdminPageAdmin(UserPermissionMixin, MultilingualModelAdmin):
 
 
 admin.site.register(ContentItem, ContentItemAdmin)
-admin.site.register(Image, ImageAdmin)
+
+if IMAGE_PREVIEW:
+    admin.site.register(Image, ImageAdminWithPreview)
+else:
+    admin.site.register(Image, ImageAdmin)
 admin.site.register(File, FileAdmin)
 admin.site.register(Page, PageAdmin)
